:sparkles: feat: new tools to manage requests to backend and organization in the client components
