+++ frontend/src/app/(root)/client/projetos/components/ComponentsManager.tsx
@@ -1,0 +1,28 @@
+import { ProjectsSkeleton } from "./ProjectSkeleton";
+import { ProjectsList } from "./Temp";
+import { NoProjects } from "./NoProjects";
+import { useQueryClient } from "@tanstack/react-query";
+
+interface Project {
+  nome: string;
+  descricao: string;
+  publico: boolean;
+}
+
+export function ComponentsManager() {
+  const queryClient = useQueryClient();
+  const queryData = queryClient.getQueryData(["get-projects"]);
+  const isLoading = !queryData;
+  const isFetched = !!queryData;
+  const projects = queryData as Project[];
+
+  if (isLoading) {
+    return <ProjectsSkeleton />;
+  }
+
+  if (isFetched && projects && projects.length > 0) {
+    return <ProjectsList proj={projects} />;
+  }
+
+  return <NoProjects />;
+}
