+++ frontend/src/app/(root)/client/projetos/components/ProjectSkeleton.tsx
@@ -1,0 +1,3 @@
+export function ProjectsSkeleton() {
+  return <div>Loading...</div>;
+}
