+++ frontend/src/app/(root)/client/projetos/stores/dialog-store.ts
@@ -1,0 +1,17 @@
+import { create } from "zustand";
+
+type DialogCreateProjetoStore = {
+  isOpen: boolean;
+  openDialog: () => void;
+  closeDialog: () => void;
+};
+
+export const useDialogManagerStore = create<DialogCreateProjetoStore>((set) => ({
+  isOpen: false,
+  openDialog: () => {
+    set({ isOpen: true });
+  },
+  closeDialog: () => {
+    set({ isOpen: false });
+  },
+}));
