+++ frontend/src/lib/react-query.ts
@@ -1,0 +1,4 @@
+"use client";
+import { QueryClient } from '@tanstack/react-query';
+
+export const queryClient = new QueryClient();
